requirements.txt
PyQt5
pyqtgraph
pyserial
numpy
pytest
PyOpenGL 
PyOpenGL_accelerate